tests: move tests to proper modules
